Update: add line comment filter module
