--- ramile/filters/sharp_comment_filter.py
@@ -1,12 +1,0 @@
-from ramile.processors import LineFilterBase
-
-
-class SharpCommentFilter(LineFilterBase):
-    """ Filters out single line comments which start with '#'
-    """
-
-    def filter(self, file, line):
-        if line.startswith('#'):
-            file.found_comment_line()
-            return line, True
-        return line, False
