+++ ramile/processors/php_processor.py
@@ -1,5 +1,5 @@
 from ramile.filters.comment_block_filter import CStyleCommentBlockFilter
-from ramile.filters.double_slash_comment_filter import DoubleSlashCommentFilter
+from ramile.filters.comment_filter import DoubleSlashCommentFilter
 from ramile.filters.html_comment_block_filter import HtmlCommentBlockFilter
 from ramile.processors import FileProcessorBase
